day2/lab 3/requirements.txt
langchain
langchain-groq
langchain-chroma
langchain-huggingface
chromadb
streamlit
python-dotenv
sentence-transformers
rank_bm25
tiktoken
